+++ src/test/scala/io/findify/s3mock/DeleteTest.scala
@@ -1,9 +1,13 @@
 package io.findify.s3mock
 
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.model._
 import com.amazonaws.services.s3.model.{AmazonS3Exception, DeleteObjectsRequest}
 
-import scala.collection.JavaConversions._
+import scala.collection.JavaConverters._
+import scala.concurrent.Await
 import scala.util.Try
+import scala.concurrent.duration._
 
 /**
   * Created by shutty on 8/11/16.
@@ -13,9 +17,9 @@ class DeleteTest extends S3MockTest {
     val s3 = fixture.client
     it should "delete a bucket" in {
       s3.createBucket("del")
-      s3.listBuckets().exists(_.getName == "del") shouldBe true
+      s3.listBuckets().asScala.exists(_.getName == "del") shouldBe true
       s3.deleteBucket("del")
-      s3.listBuckets().exists(_.getName == "del") shouldBe false
+      s3.listBuckets().asScala.exists(_.getName == "del") shouldBe false
     }
 
     it should "return 404 for non existent buckets when deleting" in {
@@ -25,9 +29,9 @@ class DeleteTest extends S3MockTest {
     it should "delete an object" in {
       s3.createBucket("delobj")
       s3.putObject("delobj", "somefile", "foo")
-      s3.listObjects("delobj", "somefile").getObjectSummaries.exists(_.getKey == "somefile") shouldBe true
+      s3.listObjects("delobj", "somefile").getObjectSummaries.asScala.exists(_.getKey == "somefile") shouldBe true
       s3.deleteObject("delobj", "somefile")
-      s3.listObjects("delobj", "somefile").getObjectSummaries.exists(_.getKey == "somefile") shouldBe false
+      s3.listObjects("delobj", "somefile").getObjectSummaries.asScala.exists(_.getKey == "somefile") shouldBe false
     }
 
     it should "return 404 for non-existent keys when deleting" in {
@@ -58,6 +62,20 @@ class DeleteTest extends S3MockTest {
       s3.putObject("delobj3", "some/path/foo2", "foo2")
       val del = s3.deleteObject("delobj3", "some/path")
       s3.listObjects("delobj3", "some/path/").getObjectSummaries.size() shouldBe 2
+    }
+
+    it should "work with aws sdk 2.0 style multi-object delete" in {
+      implicit val mat = fixture.mat
+      s3.createBucket("owntracks")
+      s3.putObject("owntracks", "data/2017-07-31/10:34.json", "foo")
+      s3.putObject("owntracks", "data/2017-07-31/16:23.json", "bar")
+      val requestData = """<?xml version="1.0" encoding="UTF-8"?><Delete xmlns="http://s3.amazonaws.com/doc/2006-03-01/"><Object><Key>data/2017-07-31/10:34.json</Key></Object><Object><Key>data/2017-07-31/16:23.json</Key></Object></Delete>"""
+      val response = Await.result(Http(fixture.system).singleRequest(HttpRequest(
+        method = HttpMethods.POST,
+        uri = s"http://localhost:${fixture.port}/owntracks?delete",
+        entity = HttpEntity(ContentType(MediaTypes.`application/xml`, HttpCharsets.`UTF-8`), requestData)
+      )), 10.seconds)
+      s3.listObjects("owntracks").getObjectSummaries.isEmpty shouldBe true
     }
   }
 }