+++ src/main/scala/io/findify/s3mock/route/ListBuckets.scala
@@ -1,6 +1,6 @@
 package io.findify.s3mock.route
 
-import akka.http.scaladsl.model.{HttpResponse, StatusCodes}
+import akka.http.scaladsl.model._
 import akka.http.scaladsl.server.Directives._
 import com.typesafe.scalalogging.LazyLogging
 import io.findify.s3mock.provider.Provider
@@ -12,7 +12,10 @@ case class ListBuckets(implicit provider:Provider) extends LazyLogging {
   def route() = get {
     complete {
       logger.debug("listing all buckets")
-      HttpResponse(StatusCodes.OK, entity = provider.listBuckets.toXML.toString)
+      HttpResponse(
+        StatusCodes.OK,
+        entity = HttpEntity(ContentType(MediaTypes.`application/xml`, HttpCharsets.`UTF-8`), provider.listBuckets.toXML.toString)
+      )
     }
   }
 }